Add files via upload
Use the Mushrooms data set for this project. Perform classification using Recursive partitioning and Regression Trees method and a simple cross validation using a 75-25 split on the data set.
This data set appears to be clean, but all of the predictors are categorical, and some have too many categories and some categories have very small frequencies (counts), so the levels need to be adjusted.